fixes search vendor detailed record view
- closes #172
- clean up npm scripts
