+++ package.json
@@ -8,28 +8,18 @@
     "start": "next start",
     "lint": "next lint",
     "typecheck": "tsc --noEmit",
-    "generate-search": "npx tsx src/scripts/generate_search_index.ts",
+    "generate-badges": "node src/scripts/generate-badges.js",
     "fetch-tweets": "ts-node -T --compilerOptions '{\"module\":\"commonjs\"}' src/scripts/fetch-tweets.ts",
     "favicon-check": "npx realfavicon check http://localhost:3000",
-    "validate-departments_json_types": "ts-node --project tsconfig.json src/scripts/update_department_json_to_type.ts",
-    "download-salary": "node src/scripts/download_publicpay_csv.js",
-    "process-salary": "npx ts-node src/scripts/process_department_salary.js",
-    "download-budgets": "src/scripts/download_budgets.sh",
-    "extract-budgets": "python src/scripts/extract_pdf_text.py",
-    "process-budgets": "npx tsx src/scripts/process_budgets.ts",
-    "download-vendors": "node src/scripts/download_vendor_transactions.js",
-    "process-vendors": "npx tsx src/scripts/process_vendors.ts",
     "test": "jest --config jest.config.js --passWithNoTests",
     "test:coverage": "npm run test -- --coverage",
     "test:coverage:full": "jest --config jest.config.js --passWithNoTests --coverage --coverageReporters=text --coverageReporters=lcov --coverageReporters=html",
     "test:unit": "TEST_MODE=mock jest --passWithNoTests --config jest.config.js",
     "test:dev": "npm run test:unit",
     "test:integration": "TEST_MODE=integration jest --config jest.config.js && npx newman run __tests__/integration/cali-doge-api-testing.postman_collection.json",
+    "test:postman": "npx newman run __tests__/integration/cali-doge-api-testing.postman_collection.json",
     "test:stage": "npm run test:integration",
-    "test:watch": "jest --config jest.config.js --watch",
-    "process-json": "npx tsx src/scripts/process_json.ts",
-    "update-supabase-types": "npx supabase gen types typescript --project-id sdopdxbeiqrktcmskplt > src/types/supabase.ts",
-    "generate-badges": "node src/scripts/generate-badges.js"
+    "test:watch": "jest --config jest.config.js --watch"
   },
   "dependencies": {
     "@radix-ui/react-select": "^2.2.2",
